move code snippets for beta features into a single file
