--- vision/cloud-client/beta/requirements.txt
@@ -1,1 +1,0 @@
-google-cloud-vision==0.27.0
