+++ vision/cloud-client/detect/resources/city.jpg
